+++ src/apps/multimedia/imx_isp.mk
@@ -10,25 +10,18 @@
 # disable the v4l_drm_test
 
 imx_isp: imx_gpu_g2d gpu_viv libdrm
-	@[ $(SOCFAMILY) != IMX ] && exit || \
+	@[ $${MACHINE:0:6} != imx8mp ] && exit || \
 	 cd $(MMDIR) && \
 	 if [ ! -d $(MMDIR)/imx_isp ]; then \
-	     wget -q $(repo_imx_isp_bin_url) -O imxisp.bin $(LOG_MUTE) && \
-	     chmod +x imxisp.bin && ./imxisp.bin --auto-accept $(LOG_MUTE) && \
+		 rm -rf imx_isp*; \
+	     $(WGET) $(repo_imx_isp_bin_url) -O imxisp.bin $(LOG_MUTE); \
+		 [ $$? -ne 0 ] && { echo "Downloading $(repo_imx_isp_bin_url) failed."; exit 1; } || \
+	     chmod +x imxisp.bin && ./imxisp.bin --auto-accept --force $(LOG_MUTE); \
 	     mv isp-imx-* imx_isp && rm -f imxisp.bin; \
 	 fi && \
-	 if [ ! -f $(DESTDIR)/usr/lib/libg2d.so ]; then \
-	     bld imx_gpu_g2d -r $(DISTROTYPE):$(DISTROVARIANT); \
+	 if [ ! -f $(DESTDIR)/usr/include/linux/dma-buf.h ]; then \
+		 bld linux-headers -r $(DISTROTYPE):$(DISTROVARIANT) -a $(DESTARCH); \
 	 fi && \
-	 if [ ! -f $(DESTDIR)/usr/lib/libOpenCL.so ]; then \
-	     bld gpu_viv -r $(DISTROTYPE):$(DISTROVARIANT); \
-	 fi && \
-	 if [ ! -f $(DESTDIR)/usr/include/xf86drm.h ]; then \
-	     bld libdrm -r $(DISTROTYPE):$(DISTROVARIANT); \
-	 fi && \
-         if [ ! -f $(DESTDIR)/usr/include/linux/dma-buf.h ]; then \
-             bld linux-headers -r $(DISTROTYPE):$(DISTROVARIANT) -a $(DESTARCH); \
-         fi && \
 	 \
 	 $(call fbprint_b,"imx_isp") && \
 	 cd imx_isp/appshell && \
